feat: 영화 상세 컴포넌트 분리 #49
