+++ src/components/movie-details/ContentRelated.tsx
@@ -1,0 +1,12 @@
+import * as S from './ContentRelated.styled'
+
+export default function ContentRelated() {
+  return (
+    <S.Container>
+      <S.Title>비슷한 콘텐츠</S.Title>
+      <S.PosterBox>
+        <S.Poster></S.Poster>
+      </S.PosterBox>
+    </S.Container>
+  )
+}
