+++ src/components/movie-search/MovieSearch.styled.ts
