+++ src/pages/index.ts
@@ -9,3 +9,5 @@ export { ReviewedList } from './ReviewedList'
 export { Bookmark } from './Bookmark'
 export { CommentDetailPage } from './CommentDetail'
 export { OtherUserProfilePage } from './OtherUserProfilePage'
+export { MovieDetails } from './MovieDetails'
+export { MovieSearch } from './MovieSearch'
